[014] Migrate StringSegment extensions into consolidated file
- Create StringSegmentExtensions.cs with IsNull, CharConversion, Reverse, CaseManipulation, CollectionOperations, Containment, and Removal regions

- Delete 10 legacy StringSegment extension files

- Remove empty subdirectories Cases/, Collections/, Contains/, Removal/

Refs: ticket-014, DECISIONS-DotExtensions-ms-extensions-collapse.md#D002
