--- DotExtensions/MsExtensions/Primitives/Cases/SegmentCaseExtensions.cs
@@ -1,65 +1,0 @@
-﻿/*
-        MIT License
-
-       Copyright (c) 2026 Alastair Lundy
-
-       Permission is hereby granted, free of charge, to any person obtaining a copy
-       of this software and associated documentation files (the "Software"), to deal
-       in the Software without restriction, including without limitation the rights
-       to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-       copies of the Software, and to permit persons to whom the Software is
-       furnished to do so, subject to the following conditions:
-
-       The above copyright notice and this permission notice shall be included in all
-       copies or substantial portions of the Software.
-
-       THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-       IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-       FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-       AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-       LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-       OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-       SOFTWARE.
-   */
-
-namespace DotExtensions.MsExtensions.Primitives;
-
-/// <summary>
-/// Provides extension methods for determining the case of <see cref="StringSegment"/> instances.
-/// </summary>
-public static class SegmentCaseExtensions
-{
-    /// <summary>
-    /// Provides extension methods for determining the case of <see cref="StringSegment"/> instances.
-    /// </summary>
-    extension(StringSegment segment)
-    {
-        /// <summary>
-        /// Returns whether this <see cref="StringSegment"/> is upper case or not.
-        /// </summary>
-        public bool IsUpperCase()
-        {
-            for (int i = 0; i < segment.Length; i++)
-            {
-                if (char.IsLower(segment[i]) || !char.IsLetter(segment[i]))
-                    return false;
-            }
-
-            return true;
-        }
-
-        /// <summary>
-        /// Returns whether a <see cref="StringSegment"/> is lower case or not.
-        /// </summary>
-        public bool IsLowerCase()
-        {
-            for (int i = 0; i < segment.Length; i++)
-            {
-                if (char.IsUpper(segment[i]) || !char.IsLetter(segment[i]))
-                    return false;
-            }
-
-            return true;
-        }
-    }
-}
