--- DotExtensions/MsExtensions/Primitives/SegmentReverseExtensions.cs
@@ -1,55 +1,0 @@
-/*
-        MIT License
-
-       Copyright (c) 2026 Alastair Lundy
-
-       Permission is hereby granted, free of charge, to any person obtaining a copy
-       of this software and associated documentation files (the "Software"), to deal
-       in the Software without restriction, including without limitation the rights
-       to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-       copies of the Software, and to permit persons to whom the Software is
-       furnished to do so, subject to the following conditions:
-
-       The above copyright notice and this permission notice shall be included in all
-       copies or substantial portions of the Software.
-
-       THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-       IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-       FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-       AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-       LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-       OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-       SOFTWARE.
-   */
-
-namespace DotExtensions.MsExtensions.Primitives;
-
-/// <summary>
-/// Provides extension methods for reversing the content of StringSegment objects.
-/// </summary>
-public static class SegmentReverseExtensions
-{
-    /// <param name="target">The StringSegment to reverse.</param>
-    extension(StringSegment target)
-    {
-        /// <summary>
-        /// Reverses the contents of the StringSegment.
-        /// </summary>
-        /// <returns>The reversed StringSegment.</returns>
-        /// <exception cref="ArgumentException">Thrown if the target is null or empty.</exception>
-        public StringSegment Reverse()
-        {
-            ArgumentException.ThrowIfNullOrEmpty(target);
-            
-            StringBuilder stringBuilder = new(capacity: target.Length);
-
-            for (int i = 0; i < target.Length; i++)
-            {
-                if (target.Length - 1 - i >= 0)
-                    stringBuilder.Append(target[target.Length - 1 - i]);
-            }
-
-            return new StringSegment(stringBuilder.ToString());
-        }
-    }
-}
